refactor: split client
