+++ commands/cfg/whoami.go
@@ -15,12 +15,6 @@ import (
 
 func WhoamiCmd() *core.Command {
 
-	type KeyInfo struct {
-		// For json printing the provenance
-		Rules   []string `json:"order"`
-		UsedSrc string   `json:"used"`
-	}
-
 	const (
 		FlagProvenance      = "provenance"
 		FlagProvenanceShort = "p"
