+++ internal/client/client.go
@@ -9,71 +9,13 @@ import (
 	"github.com/ionos-cloud/ionosctl/v6/internal/die"
 	"github.com/ionos-cloud/ionosctl/v6/pkg/config"
 	"github.com/ionos-cloud/ionosctl/v6/pkg/constants"
-	sdkgoauth "github.com/ionos-cloud/sdk-go-auth"
-	certmanager "github.com/ionos-cloud/sdk-go-cert-manager"
-	registry "github.com/ionos-cloud/sdk-go-container-registry"
-	dataplatform "github.com/ionos-cloud/sdk-go-dataplatform"
-	mongo "github.com/ionos-cloud/sdk-go-dbaas-mongo"
-	postgres "github.com/ionos-cloud/sdk-go-dbaas-postgres"
-	cloudv6 "github.com/ionos-cloud/sdk-go/v6"
 	"github.com/spf13/viper"
 )
-
-type Client struct {
-	UsedLayer Layer // i.e. which auth layer are we using. Flags / Env Vars / Config File
-
-	CloudClient        *cloudv6.APIClient
-	AuthClient         *sdkgoauth.APIClient
-	CertManagerClient  *certmanager.APIClient
-	PostgresClient     *postgres.APIClient
-	MongoClient        *mongo.APIClient
-	DataplatformClient *dataplatform.APIClient
-	RegistryClient     *registry.APIClient
-}
-
-func appendUserAgent(userAgent string) string {
-	return fmt.Sprintf("%v_%v", viper.GetString(constants.CLIHttpUserAgent), userAgent)
-}
-
-func newClient(name, pwd, token, hostUrl string) *Client {
-	clientConfig := cloudv6.NewConfiguration(name, pwd, token, hostUrl)
-	clientConfig.UserAgent = fmt.Sprintf("%v_%v", viper.GetString(constants.CLIHttpUserAgent), clientConfig.UserAgent)
-	// Set Depth Query Parameter globally
-	clientConfig.SetDepth(1)
-
-	authConfig := sdkgoauth.NewConfiguration(name, pwd, token, hostUrl)
-	authConfig.UserAgent = appendUserAgent(authConfig.UserAgent)
-
-	certManagerConfig := certmanager.NewConfiguration(name, pwd, token, hostUrl)
-	certManagerConfig.UserAgent = appendUserAgent(certManagerConfig.UserAgent)
-
-	postgresConfig := postgres.NewConfiguration(name, pwd, token, hostUrl)
-	postgresConfig.UserAgent = appendUserAgent(postgresConfig.UserAgent)
-
-	mongoConfig := mongo.NewConfiguration(name, pwd, token, hostUrl)
-	mongoConfig.UserAgent = appendUserAgent(mongoConfig.UserAgent)
-
-	dpConfig := dataplatform.NewConfiguration(name, pwd, token, hostUrl)
-	dpConfig.UserAgent = appendUserAgent(dpConfig.UserAgent)
-
-	registryConfig := registry.NewConfiguration(name, pwd, token, hostUrl)
-	registryConfig.UserAgent = appendUserAgent(registryConfig.UserAgent)
-
-	return &Client{
-		CloudClient:        cloudv6.NewAPIClient(clientConfig),
-		AuthClient:         sdkgoauth.NewAPIClient(authConfig),
-		CertManagerClient:  certmanager.NewAPIClient(certManagerConfig),
-		PostgresClient:     postgres.NewAPIClient(postgresConfig),
-		MongoClient:        mongo.NewAPIClient(mongoConfig),
-		DataplatformClient: dataplatform.NewAPIClient(dpConfig),
-		RegistryClient:     registry.NewAPIClient(registryConfig),
-	}
-}
 
 var once sync.Once
 var instance *Client
 
-func getFirstValidSource(layers []Layer) (values map[string]string, usedLayer Layer) {
+func selectAuthLayer(layers []Layer) (values map[string]string, usedLayer Layer, err error) {
 	for _, layer := range layers {
 		token := viper.GetString(layer.TokenKey)
 		username := viper.GetString(layer.UsernameKey)
@@ -85,23 +27,10 @@ func getFirstValidSource(layers []Layer) (values map[string]string, usedLayer La
 				"token":    token,
 				"username": username,
 				"password": password,
-			}, layer
+			}, layer, nil
 		}
 	}
-	return nil, Layer{}
-}
-
-type Layer struct {
-	TokenKey    string
-	UsernameKey string
-	PasswordKey string
-	Help        string
-}
-
-var ConfigurationPriorityRules = []Layer{
-	{constants.ArgToken, "", "", fmt.Sprintf("Global Flags (--%s)", constants.ArgToken)},
-	{constants.EnvToken, constants.EnvUsername, constants.EnvPassword, fmt.Sprintf("Environment Variables (%s, %s, %s)", constants.EnvToken, constants.EnvUsername, constants.EnvPassword)},
-	{constants.CfgToken, constants.CfgUsername, constants.CfgPassword, fmt.Sprintf("Config file settings (%s, %s, %s)", constants.CfgToken, constants.CfgUsername, constants.CfgPassword)}, // Note: Username & Password are no longer generated in cfg file by `ionosctl login`, however we will keep this for backward compatibility.
+	return nil, Layer{}, fmt.Errorf("none of the layers provided a value for either token or ")
 }
 
 // Get a client and possibly fail. Uses viper to get the credentials and API URL.
@@ -128,7 +57,7 @@ func Get() (*Client, error) {
 
 		viper.AutomaticEnv()
 
-		values, usedLayer := getFirstValidSource(ConfigurationPriorityRules)
+		values, usedLayer, err := selectAuthLayer(ConfigurationPriorityRules)
 		if values == nil {
 			getClientErr = errors.Join(getClientErr, fmt.Errorf("not logged in: use either environment variables %s or %s and %s, or use `ionosctl login`", constants.EnvToken, constants.EnvUsername, constants.EnvPassword))
 			return
@@ -189,9 +118,4 @@ func (c *Client) TestCreds() error {
 	}
 
 	return nil
-}
-
-// IsTokenAuth returns true if a token is being used for authentication. Otherwise, username & password were used.
-func (c *Client) IsTokenAuth() bool {
-	return c.CloudClient.GetConfig().Token != ""
 }
