+++ internal/client/model.go
@@ -1,0 +1,87 @@
+package client
+
+import (
+	"fmt"
+
+	"github.com/ionos-cloud/ionosctl/v6/pkg/constants"
+	sdkgoauth "github.com/ionos-cloud/sdk-go-auth"
+	certmanager "github.com/ionos-cloud/sdk-go-cert-manager"
+	registry "github.com/ionos-cloud/sdk-go-container-registry"
+	dataplatform "github.com/ionos-cloud/sdk-go-dataplatform"
+	mongo "github.com/ionos-cloud/sdk-go-dbaas-mongo"
+	postgres "github.com/ionos-cloud/sdk-go-dbaas-postgres"
+	cloudv6 "github.com/ionos-cloud/sdk-go/v6"
+	"github.com/spf13/viper"
+)
+
+var ConfigurationPriorityRules = []Layer{
+	{constants.ArgToken, "", "", fmt.Sprintf("Global Flags (--%s)", constants.ArgToken)},
+	{constants.EnvToken, constants.EnvUsername, constants.EnvPassword, fmt.Sprintf("Environment Variables (%s, %s, %s)", constants.EnvToken, constants.EnvUsername, constants.EnvPassword)},
+	{constants.CfgToken, constants.CfgUsername, constants.CfgPassword, fmt.Sprintf("Config file settings (%s, %s, %s)", constants.CfgToken, constants.CfgUsername, constants.CfgPassword)}, // Note: Username & Password are no longer generated in cfg file by `ionosctl login`, however we will keep this for backward compatibility.
+}
+
+type Client struct {
+	UsedLayer Layer // i.e. which auth layer are we using. Flags / Env Vars / Config File
+
+	CloudClient        *cloudv6.APIClient
+	AuthClient         *sdkgoauth.APIClient
+	CertManagerClient  *certmanager.APIClient
+	PostgresClient     *postgres.APIClient
+	MongoClient        *mongo.APIClient
+	DataplatformClient *dataplatform.APIClient
+	RegistryClient     *registry.APIClient
+}
+
+// Layer represents an authentication layer. E.g., flags, env vars, config file.
+type Layer struct {
+	TokenKey    string
+	UsernameKey string
+	PasswordKey string
+	Help        string
+}
+
+// IsTokenAuth returns true if a token is being used for authentication. Otherwise, username & password were used.
+func (c *Client) IsTokenAuth() bool {
+	return c.CloudClient.GetConfig().Token != ""
+}
+
+// -- private funcs --
+
+func appendUserAgent(userAgent string) string {
+	return fmt.Sprintf("%v_%v", viper.GetString(constants.CLIHttpUserAgent), userAgent)
+}
+
+func newClient(name, pwd, token, hostUrl string) *Client {
+	clientConfig := cloudv6.NewConfiguration(name, pwd, token, hostUrl)
+	clientConfig.UserAgent = fmt.Sprintf("%v_%v", viper.GetString(constants.CLIHttpUserAgent), clientConfig.UserAgent)
+	// Set Depth Query Parameter globally
+	clientConfig.SetDepth(1)
+
+	authConfig := sdkgoauth.NewConfiguration(name, pwd, token, hostUrl)
+	authConfig.UserAgent = appendUserAgent(authConfig.UserAgent)
+
+	certManagerConfig := certmanager.NewConfiguration(name, pwd, token, hostUrl)
+	certManagerConfig.UserAgent = appendUserAgent(certManagerConfig.UserAgent)
+
+	postgresConfig := postgres.NewConfiguration(name, pwd, token, hostUrl)
+	postgresConfig.UserAgent = appendUserAgent(postgresConfig.UserAgent)
+
+	mongoConfig := mongo.NewConfiguration(name, pwd, token, hostUrl)
+	mongoConfig.UserAgent = appendUserAgent(mongoConfig.UserAgent)
+
+	dpConfig := dataplatform.NewConfiguration(name, pwd, token, hostUrl)
+	dpConfig.UserAgent = appendUserAgent(dpConfig.UserAgent)
+
+	registryConfig := registry.NewConfiguration(name, pwd, token, hostUrl)
+	registryConfig.UserAgent = appendUserAgent(registryConfig.UserAgent)
+
+	return &Client{
+		CloudClient:        cloudv6.NewAPIClient(clientConfig),
+		AuthClient:         sdkgoauth.NewAPIClient(authConfig),
+		CertManagerClient:  certmanager.NewAPIClient(certManagerConfig),
+		PostgresClient:     postgres.NewAPIClient(postgresConfig),
+		MongoClient:        mongo.NewAPIClient(mongoConfig),
+		DataplatformClient: dataplatform.NewAPIClient(dpConfig),
+		RegistryClient:     registry.NewAPIClient(registryConfig),
+	}
+}
